Moves Biome-related classes and enums to Biomes namespace.
