+++ ParquetClassLibrary/ParquetClassLibrary.csproj
@@ -28,6 +28,7 @@
     <Folder Include="Utilities\" />
     <Folder Include="Quests\" />
     <Folder Include="Items\" />
+    <Folder Include="Biomes\" />
   </ItemGroup>
 
   <ItemGroup>
